requirements.txt
python-dotenv
requests
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
beautifulsoup4Flask
